==> analysis/01-convert-ffl-lists.ipynb <==
# Convert excel or txt files to csv
The complete list of federal firearms licensees is posted monthly on the [ATF website](https://www.atf.gov/firearms/listing-federal-firearms-licensees). It is usually available as either a txt file or an excel file. To make batch processing easier, we converted all of these files into csvs. 

import pandas as pd

# For xlsm files, add engine="openpyxl"
# For 2012, add sheet_name="FFL January 2012"
# df = pd.read_excel('./_workspace/gun-dealers-2023-analysis/data/source/atf-ffl-list/2013/2013 FFL List.xlsx', dtype=str, skiprows=[1]) 
df = pd.read_csv('../../data/source/atf-ffl-list/2023/1123-ffl-list.txt', sep="\t", skiprows=[1], dtype=str)

df.head()

df.to_csv('../../data/source/atf-ffl-list/2023/1123-ffl-list.csv', index=False)
