==> analysis/02-create-county-reference-files.ipynb <==
# Create county information dataset

import pandas as pd
import geopandas as gpd

## Load and clean county shapefile

counties_gdf = gpd.read_file("../../data/source/census/cb_2018_us_county_500k/cb_2018_us_county_500k.shp")

# Set to NAD 1983 Albers North America - https://epsg.io/102008
counties_gdf = counties_gdf.to_crs("ESRI:102008")

# Convert square meter to square miles - https://www.metric-conversions.org/area/square-meters-to-square-miles.htm
counties_gdf["area"] = counties_gdf.area * 0.00000038610215855

counties_gdf.rename(columns={"GEOID":"fips"}, inplace=True)

counties_gdf.sample()

counties_gdf_trim = counties_gdf[["fips","area"]]

counties_gdf_trim.sample()

## Find neighbor counties

### Create the neighbors nested JSON file

neighbors = {}

for county in counties_gdf.itertuples():
    neighbor_filter = counties_gdf[counties_gdf.touches(counties_gdf.geometry)]
    neighbors[counties_gdf.fips] = neighbor_filter["geoid"].tolist()

with open("neighbor-counties.json", "w") as outfile:
    json.dump(neighbors, outfile)

### Create the long neighbors CSV file

# Read the list of neighbors
neighbors = open("../../data/handmade/neighbor-counties.json")
neighbors = json.load(neighbors)

neighbors = pd.DataFrame(pd.concat({k: pd.Series(v) for k, v in neighbors.items()})).reset_index()

neighbors.drop(['level_1'], axis=1, inplace=True)
neighbors.columns = ["fips","neighbor_fips"]

neighbors.sample()

# neighbors.to_csv("../../data/handmade/neighbor-counties.csv", index=False)

## Load and clean state and county FIPS code files

state_fips = pd.read_csv('../../data/source/census/state_fips.txt', sep="|", dtype=str)

state_fips.columns = state_fips.columns.str.lower()

state_fips.sample()

county_fips = pd.read_csv('../../data/source/census/county_fips.txt', sep="|", dtype=str)

county_fips.columns = county_fips.columns.str.lower()

# Create county FIPS code
county_fips["fips"] = county_fips["statefp"] + county_fips["countyfp"]

county_fips.sample()

# Filter states
county_fips = county_fips.loc[~county_fips["state"].isin(['AS', 'GU', 'MP', 'PR', 'UM', 'VI'])]

county_fips.state.unique()

counties = county_fips[["state","statefp","countyname","countyfp","fips"]].merge(counties_gdf_trim, on="fips", how="left")

counties.sample()

## Save the county reference file

counties.to_csv("../../data/processed/counties-reference.csv", index=False)
