==> analysis/04-geocode-ffls.ipynb <==
# Geocode January 2023 FFLs
This data will be useful for calculating the driving distances

df = pd.read_csv('../../data/processed/atf-ffl-list/2023-ffl-list-clean.csv')

df.sample()

len(df)

# Create full address column
df["original_address"] = df["premise_street"] + ", " + df["premise_city"] + ", " + df["premise_state"] + ", "  + df["premise_zip_code"]

batch = df[["original_address"]].copy()

batch = batch.drop_duplicates().reset_index(drop=True)

batch['longitude'] = ""
batch['latitude'] = ""
batch['clean_address'] = ""

batch.head()

# Specify which API key to use
MAPS_KEY = os.getenv('MAPS_KEY')

url = 'https://maps.googleapis.com/maps/api/geocode/json?'

for x in range(len(batch)):
    place = batch['original_address'][x]
    print(x, ":", place)
    try:
        geocode_result = requests.get(url + 'address=' + place + '&key=' + MAPS_KEY).json()
        batch['longitude'][x] = geocode_result.get('results')[0].get('geometry').get('location').get('lng')
        batch['latitude'][x] = geocode_result.get('results')[0].get('geometry').get('location').get('lat')
        batch['clean_address'][x] = geocode_result.get('results')[0].get('formatted_address')
    except:
        batch['longitude'][x] = ""
        batch['latitude'][x] = ""
        batch['clean_address'][x] = ""

batch

batch.to_csv("./_workspace/gun-dealers-2023-analysis/data/processed/dealers-geocoded-2023.csv", index=False)

# These may need to be fixed manually
error = batch.loc[batch["clean_address"] == ""]

error.to_csv('./_workspace/gun-dealers-2023-analysis/data/processed/atf-ffl-list/geocoding_error_2023.csv', index=False)
