==> analysis/05-clean-cdc-data.ipynb <==
# Clean CDC Wonder data

import pandas as pd
import numpy as np

# These are the years of CDC data that we have
years = [
    '2010', '2011', '2012', '2013',
    '2014', '2015', '2016', '2017', 
    '2018', '2019', '2020', '2021'
        ]

for year in years:
    df_deaths = pd.read_csv('./_workspace/gun-dealers-2023-analysis/data/processed/cdc-data/mult'+year+'_grouped_counties.csv', dtype={"geoid":'str'})
    df_deaths.rename(columns={"geoid":"fips"}, inplace=True)
    population = pd.read_csv('./_workspace/gun-dealers-2023-analysis/data/processed/census/acs5_'+year+'_population_counties.csv', dtype={"geoid":'str'}, usecols=["geoid","universe"])
    population.rename(columns={"geoid":"fips","universe":"population"}, inplace=True)
    df_deaths = df_deaths.merge(population, on="fips", how="left")
    df_deaths["total_rate"] = (df_deaths["deaths"] / df_deaths["population"]) *100000
    df_deaths.to_csv('./_workspace/gun-dealers-2023-analysis/data/processed/cdc-data/mult'+year+'_grouped_rates_counties.csv', index=False)

gun_hom = df_deaths.loc[df_deaths["type_manner"] == "firearm-homicide"].sort_values("deaths", ascending=False).reset_index(drop=True)

gun_hom.iloc[0:100]

gun_hom.iloc[0:100]["deaths"].sum()

top_100 = gun_hom.iloc[0:100]["fips"]

top_200 = gun_hom.iloc[0:200]["fips"]

gun_hom.iloc[0:100]["deaths"].sum() / gun_hom["deaths"].sum()

gun_hom.iloc[0:51]["deaths"].sum() / gun_hom["deaths"].sum()

top_50 = gun_hom.iloc[0:51]["fips"]

gun_hom.head(51)

top_50.to_csv('./_workspace/gun-dealers-2023-analysis/data/processed/top_50_gun_homicide_deaths_2021.csv', index=False)
top_100.to_csv('./_workspace/gun-dealers-2023-analysis/data/processed/top_100_gun_homicide_deaths_2021.csv', index=False)
top_200.to_csv('./_workspace/gun-dealers-2023-analysis/data/processed/top_200_gun_homicide_deaths_2021.csv', index=False)
