==> analysis/06-clean-ca-crime-gun-data.ipynb <==
# Clean 2023 CA crime gun report data
The table at the end of [this report](https://oag.ca.gov/system/files/attachments/press-docs/AB%201191%20Crime%20Gun%20Report.pdf) contains data about the number of guns sold and the number that were later recovered by California law enforcement (crime guns) for all California-based dealers with at least one crime gun. We extracted the table using tabula.

import pandas as pd
import altair as alt
import os

## Clean up tabula extraction and save raw data

ca_crime_guns = pd.read_csv("../../data/source/tabula-AB 1191 Crime Gun Report -Dealer Information.csv",
                            dtype=str)

ca_crime_guns = ca_crime_guns.replace("\r"," ",regex=True)

ca_crime_guns.columns = ca_crime_guns.columns.str.replace("\r"," ",regex=True)

ca_crime_guns["Dealer Name"] = ca_crime_guns["Dealer Name"].str.replace("’","",regex=True)
ca_crime_guns["Crime Guns"] = ca_crime_guns["Crime Guns"].str.replace(",","")
ca_crime_guns["Total Firearms Sold"] = ca_crime_guns["Total Firearms Sold"].str.replace(",","")
ca_crime_guns["Total Firearms Sold"] = ca_crime_guns["Total Firearms Sold"].str.replace("%","")
ca_crime_guns["Percentage Crime Guns"] = ca_crime_guns["Percentage Crime Guns"].str.replace("%","")

ca_crime_guns["Crime Guns"] = pd.to_numeric(ca_crime_guns["Crime Guns"])
ca_crime_guns["Total Firearms Sold"] = pd.to_numeric(ca_crime_guns["Total Firearms Sold"])
ca_crime_guns["Percentage Crime Guns"] = pd.to_numeric(ca_crime_guns["Percentage Crime Guns"])

len(ca_crime_guns)

ca_crime_guns.columns = ["dealer_name","dealer_status","street","zip_code","city","crime_guns","total_firearms_sold","percentage_crime_guns"]

ca_crime_guns.head()

ca_crime_guns.to_csv("../../data/processed/ca_ab_1191_crime_gun_report.csv", index=False)
