==> analysis/07-generate-isochrones.ipynb <==
# Generate isochrones by state
To determine the driving distance from each dealer, we used the open route service's isochrone feature. It is documented here. We did each state individually to speed up processing time. 

import pandas as pd
import requests
import json

# docker compose up
# https://download.geofabrik.de/north-america/us.html

ffls = pd.read_csv("../../data/processed/atf-ffl-list/ffls_2023.csv")

# ffls.head()

state_ffls = ffls[ffls["state_name"] == "New Mexico"]

len(state_ffls)

# state_ffls.iloc[0:10]

target = ['NM']
abbr = 'nm'
state = 'new-mexico'
distance = 600 # 10 minutes (600), 20 minutes (1200), 30 minutes (1800)
mvmt_type = 'driving'

# Create empty feature collection
collection = {'type': 'FeatureCollection', 'features': []}
no_iso = pd.DataFrame()
i = 0

for row in state_ffls.itertuples(index=False):
    print(i, row.original_address)
    try:
        # Set variables
        request_locations = [[row.longitude,row.latitude]]
        request_range = [distance]
        body = {
            "locations": request_locations,
            "range": request_range,
            "location_type": "start",
            "range_type": "time",
        }
        headers = {
            'Accept': 'application/json, application/geo+json, application/gpx+xml, img/png; charset=utf-8',
            'Content-Type': 'application/json; charset=utf-8'
        }
        # Call the request
        # call = requests.post('http://localhost:8080/ors/v2/isochrones/foot-walking', json=body, headers=headers)
        call = requests.post('http://localhost:8080/ors/v2/isochrones/driving-car', json=body, headers=headers)
        # Pull out the feature
        data = json.loads(call.text)
        if "features" in data:
            feature = data["features"][0]
            # Add the FFL id
            feature["properties"]["original_address"] = row.original_address
            # Append to the feature collection
            collection["features"].append(feature)
        else:
            no_iso = pd.concat([no_iso, state_ffls.iloc[[i]]])
    except:
        no_iso = pd.concat([no_iso, state_ffls.iloc[[i]]])
    
    i = i + 1

len(collection.get('features'))

with open('../../data/processed/isochrones/'+abbr+'_'+str(distance)+'_'+mvmt_type+'.geojson', 'w') as f:
    json.dump(collection, f)
