==> analysis/08-merge-isochrones.ipynb <==
import pandas as pd
import geopandas as gpd
import shapely

states = [
    "ak","al","ar","az","co",
    "ct","dc","de","fl","ga",
    "hi","ia","id","il","in",
    "ks","ky","la","ma","md",
    "me","mi","mn","mo","ms",
    "mt","nc","nd","ne","nh",
    "nj","nm","nor-ca","nv",
    "ny","oh","ok","or","pa",
    "ri","sc","sd","so-ca",
    "tn","tx","ut","va","vt",
    "wa","wi","wv","wy"
]

for state in states:
    print(state)
    iso = gpd.read_file("../../data/processed/isochrones/raw/"+state+"_600_driving.geojson")
    iso['geometry'] = shapely.set_precision(iso['geometry'].values, 1e-6)
    # dissolved = iso.dissolve()
    # dissolved.plot()
    # dissolved.to_file("../../data/processed/isochrones/dissolved/"+state+"_600_driving_dissolved.geojson", driver='GeoJSON')
    union = gpd.overlay(iso, iso, how='union', keep_geom_type=False)
    union["geometry_str"] = union.geometry.astype(str)
    overlaps = union.dissolve(
         by="geometry_str",
         aggfunc={
             "original_address_2": "count"
         },
     )
    overlaps_clean = overlaps.reset_index(drop=True)
    overlaps_clean.columns = ["geometry","count"]
    max = overlaps_clean["count"].max()
    if max <= 20:
        max = 21
    overlaps_clean['bin'] = pd.cut(overlaps_clean['count'], [0, 1, 5, 10, 20, max], labels=['1', '1-5', '5-10', '10-20', '20+'])
    overlaps_binned = overlaps_clean.dissolve(by='bin')
    overlaps_binned = overlaps_binned.reset_index()
    overlaps_binned['bin'] = overlaps_binned['bin'].astype(str)
    overlaps_binned.to_file("../../data/processed/isochrones/"+state+"_600_driving_binned.geojson", driver='GeoJSON')

### Results from QGIS analysis

populaton_within_10_minutes = 302546129.13564754

# The zonal statistics total population is higher than the 333,287,550 Census number
us_population = 343326438.89526474 

populaton_within_10_minutes / us_population
