==> analysis/13-ca-crime-gun-analysis.ipynb <==
import pandas as pd
import altair as alt
import os

## Which dealers sell the most guns?

ca_crime_guns = pd.read_csv("data/processed/ca_ab_1191_crime_gun_report.csv")

ca_crime_guns.sort_values("total_firearms_sold", ascending=False).head()

# Total guns sold
ca_crime_guns["total_firearms_sold"].sum()

## Which dealers have a higher than average percentage of crime guns sold?

ca_crime_guns["percentage_crime_guns"].mean()

ca_crime_guns.iloc[0:82]["crime_guns"].sum() / ca_crime_guns["crime_guns"].sum()

top_sellers = ca_crime_guns.iloc[0:82]
