==> analysis/15-calculate-homicide-rates.ipynb <==
# Calculate homicide rate

import pandas as pd

hom_99_20 = pd.read_csv('./_workspace/gun-dealers-2023-analysis/data/processed/cdc-data/Underlying Cause of Death, 1999-2020.txt', sep="\t", usecols=["Year","Deaths","Population"])
hom_18_21 = pd.read_csv('./_workspace/gun-dealers-2023-analysis/data/processed/cdc-data/Underlying Cause of Death, 2018-2021, Single Race.txt', sep="\t", usecols=["Year","Deaths","Population"])

hom_99_20 = hom_99_20[hom_99_20["Year"] >= 2014]
hom_18_21 = hom_18_21[hom_18_21["Year"] == 2021]

hom_18_21

hom_14_21 = pd.concat([hom_99_20,hom_18_21])

hom_14_21["total_rate"] = (hom_14_21["Deaths"] / hom_14_21["Population"]) * 100000

hom_14_21

# Firearm homicide rate for all analysis years
(hom_14_21["Deaths"].sum() / hom_14_21["Population"].sum()) * 100000

# Firearm homicide count for each of the analysis years
hom_14_21["Deaths"].sum() / 8

# Firearm homicide rate for 2021
rate_21 = hom_18_21["Deaths"].sum() / hom_18_21["Population"].sum() * 100000

rate_21
